--- a/src/runtime/runtime.h
+++ b/src/runtime/runtime.h
@@ -183,6 +183,7 @@ typedef closure_type(buffer_handler, status, buffer);
 typedef closure_type(connection_handler, buffer_handler, buffer_handler);
 typedef closure_type(io_status_handler, void, status, bytes);
 typedef closure_type(block_io, void, void *, range, status_handler);
+typedef closure_type(block_sync, void, thunk);
 
 #include <sg.h>