--- a/src/runtime/runtime.h
+++ b/src/runtime/runtime.h
@@ -29,6 +29,7 @@ void console_write(const char *s, bytes count);
 void print_u64(u64 s);
 
 void halt(char *format, ...) __attribute__((noreturn));
+void kernel_shutdown(int status) __attribute__((noreturn));
 void vm_exit(u8 code) __attribute__((noreturn));
 void print_stack_from_here();