… stage3/Makefile
… test/runtime/write.c
… test/runtime/write.manifest
… test/runtime/write_contents/hello