This update to the pagecache applies write completions immediately upon posting the write to the page buffer, resulting in improved performance as file writes no longer need to wait for I/O completion to finish. Spurred by scheduling changes needed to implement file-backed page faults in the kernel, this also allows pagecache storage I/O completions to be safely applied outside of the scope of the kernel lock (upcoming PR). Pagecache read/write completions are now queued up to be applied in a service routine that runs on runqueue (which will remain under the kernel lock).

A sync function has been added to the cache to allow flushing of pending writes on fsync(2), fdatasync(2), newly-added sync(2) and syncfs(2), and on system shutdown.

A simple bulk write performance test has been added to the write runtime test. Comparison with master shows a > 40% improvement when writing, syncing and reading (from cache) a 20M file (with virtio-scsi). Other file sizes can be specified on the command line with the '-s' option.

master:
```
bulk write test, size 20480 KB:
   write   0.079221725s
   fsync   0.000997913s
    read   0.020217010s
   total   0.100436648s (203909 KB/s)
write test passed
```

async writes:
```
bulk write test, size 20480 KB:
   write   0.048795145s
   fsync   0.000968155s
    read   0.020555042s
   total   0.070318342s (291246 KB/s)
write test passed
```

the write test also exposes serious performance degradations with virito-blk and ata drivers:

with STORAGE=virtio-blk:
```
bulk write test, size 20480 KB:
   write   4.313854939s
   fsync   0.062707746s
    read   0.022653938s
   total   4.399216623s (4655 KB/s)
write test passed
```

with STORAGE=ide:
```
bulk write test, size 20480 KB:
   write   33.672703707s
   fsync   0.000907161s
    read   0.020189055s
   total   33.693799923s (607 KB/s)
write test passed
```

Resolves #1128 and #1166